+++ src/server/devices/models/index.ts
@@ -1,2 +1,3 @@
 export * from './WindowPosition';
 export * from './MagnetPosition';
+export * from './phaseState';